refactoring client connection and lobby managers, messages
