+++ src/api/mod.rs
@@ -1,9 +1,8 @@
+pub mod message;
 mod lobby_api;
 mod client_connection_api;
 
 use actix_web::{web, HttpResponse};
-pub use lobby_api::lobby_manager::{LobbyManager, LobbyId};
-pub use client_connection_api::client_connection_manager::{ClientConnectionManager, ClientConnectionId};
 
 pub fn config(cfg: &mut web::ServiceConfig) {
   cfg.service(
